+++ absl/CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com).
 
 ### Added
 
+*   (testing) flagsaver overrides can now be specified in terms of FlagHolder.
 *   (testing) `parameterized.product`: Allows testing a method over cartesian
     product of parameters values.